requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
python-dotenv==1.0.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
aiohttp==3.11.18
